Move maps to separate file
